enable reply zero in assistant setup
